app/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]
python-jose
pydantic
pydantic-settings
python-multipart
email-validator